+++ packages/speculator/CHANGELOG.md
@@ -7,3 +7,4 @@
 - 871ca82: refactor reference handling and introducing basic linting
 - 5bac7e2: extend ast to render w3c conform spec header
 - e55f102: Intoroducting from automatic confrmance and biblio sections, + additional lint rules
+- Add environment injection and refactor buildWorkspaces API